+++ apps/web/src/app/(onboarding)/onboarding-prompt-engineering/generate-dataset/_components/GenerateDataset/index.tsx
@@ -7,7 +7,7 @@ import { Suspense, useCallback } from 'react'
 import { BlocksEditorPlaceholder } from '$/components/BlocksEditor'
 import { OnboardingEditor } from '../../../_components/OnboardingEditor'
 import { TableSkeleton } from '@latitude-data/web-ui/molecules/TableSkeleton'
-import { usePromptEngineeringOnboarding } from '$/app/(onboarding)/onboarding-prompt-engineering/datasetOnboarding'
+import { usePromptEngineeringOnboarding } from '$/app/(onboarding)/onboarding-prompt-engineering/usePromptEngineeringOnboarding'
 import { ROUTES } from '$/services/routes'
 import { useNavigate } from '$/hooks/useNavigate'
