+++ apps/web/src/app/(onboarding)/onboarding-prompt-engineering/run-experiment/_components/RunExperiment/index.tsx
@@ -14,7 +14,7 @@ import SimpleDatasetTable from '../../../_components/SimpleDatasetTable'
 import { useExperiments } from '$/stores/experiments'
 import useDatasets from '$/stores/datasets'
 import { envClient } from '$/envClient'
-import { usePromptEngineeringOnboarding } from '$/app/(onboarding)/onboarding-prompt-engineering/datasetOnboarding'
+import { usePromptEngineeringOnboarding } from '$/app/(onboarding)/onboarding-prompt-engineering/usePromptEngineeringOnboarding'
 import useWorkspaceOnboarding from '$/stores/workspaceOnboarding'
 import useLatitudeAction from '$/hooks/useLatitudeAction'
 import { publishEventAction } from '$/actions/events/publishEventAction'
